requirements.txt
numpy>=1.24.0
pygame-ce>=2.3.0
matplotlib>=3.7.0
pandas>=2.0.0
scipy>=1.10.0
seaborn>=0.12.0
tqdm>=4.60.0